improved webpack configs
