+++ webpack.config.dev.js
@@ -1,63 +1,18 @@
-const CopyWebpackPlugin = require('copy-webpack-plugin')
-const HTMLWebpackPlugin = require('html-webpack-plugin')
+const { merge } = require('webpack-merge');
 const path = require("path");
 
-const babelPresetEnv = require.resolve('@babel/preset-env');
-const babelPresetReact = require.resolve('@babel/preset-react');
+const common = require('./webpack.config.common.js');
 
-module.exports = {
+const devConfig = {
     mode: 'development',
     devServer: {
-        contentBase: 'devs',
+        contentBase: 'dev',
         port: 3000
     },
     devtool: 'inline-source-map',
     output: {
-        path: path.resolve(__dirname, 'devs')
+        path: path.resolve(__dirname, 'dev')
     },
-    resolve: {
-        alias: {
-            vue: 'vue/dist/vue.js',
-        }
-    },
-    plugins: [
-        new CopyWebpackPlugin({
-            patterns: [
-                {
-                    from: 'build/assets',
-                    to: 'assets'
+};
 
-                },
-                {
-                    from: 'build/styles',
-                    to: 'styles'
-
-                }
-            ]
-        }),
-        new HTMLWebpackPlugin({
-            template: 'build/index.html',
-            filename: 'index.html'
-        })
-    ],
-    module: {
-        rules: [
-            { test: /\.scss?$/, loaders: ['style-loader', 'css-loader', 'sass-loader'] },
-            { test: /\.css?$/, loaders: ['style-loader', 'css-loader', 'sass-loader'] },
-            {
-                test: /\.(js|jsx)$/,
-                exclude: /node_modules/,
-                use: {
-                    loader: 'babel-loader',
-                    options: {
-                        presets: [
-                            babelPresetEnv,
-                            babelPresetReact
-                        ],
-                        plugins: []
-                    }
-                }
-            }
-        ]
-    },
-}
+module.exports = merge(common, devConfig);
